+++ AWS/CHANGELOG.md
@@ -7,11 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## 2024-03-22 - 1.30.9
+## 2024-03-29 - 1.30.9
 
 ### Changed
 
-- Use chunk size from environment variable
+- Update `sekoia-automation-sdk` to the latest version (1.12.2) to fix
+  support to intake batch URL.
 
 ## 2024-03-22 - 1.30.8